[refac]: moved class "ODDReader" to own file
